+++ CHANGELOG.md
@@ -8,7 +8,6 @@
 
 ## Queue
 
-- [ ] Short flags
 - [ ] Display manager
 - [ ] nvim.desktop
 - [ ] Theme for Win+D launcher
@@ -18,6 +17,7 @@
     - [ ] nvim monokai theme is broken
 - Apps:
     - [ ] autoproxy
+- [x] Short flags
 - [x] --no-update
 - [x] move example.lua to example/ anyway, mess with package.path or something
 - [x] blueman notifications are annoying af